add python3 install script

+++ init_env.sh
@@ -1,9 +1,10 @@
 #!/bin/bash
 
-if [ $# -ne 1 ];then
-    echo "usage: sh init_env.sh work/nginx/ss"
+function usage()
+{
+    echo "usage: sh init_env.sh work/nginx/ss/python3"
     exit 0
-fi
+}
 
 function init_work()
 {
@@ -32,11 +33,21 @@ function init_work()
 function deploy_nginx()
 {
     cd
+    #check existance of nginx
+    check_str=$(whereis nginx | awk -F ': ' '{print $2}')
+    echo "check_str:$check_str"
+    if [ "$check_str" != "" ];then
+        echo "nginx exist."
+        whereis nginx
+        nginx -V
+        return
+    fi
+    #close firewall
     firewall-cmd --state
     systemctl stop firewalld.service
     systemctl disable firewalld.service
     firewall-cmd --state
-
+    #start install nginx
     yum install gcc-c++
     yum install -y pcre pcre-devel
     yum install -y zlib zlib-devel
@@ -46,6 +57,7 @@ function deploy_nginx()
     ./configure
     make
     make install
+    #create soft link
     ln -s /usr/local/nginx/sbin/nginx /usr/local/bin/nginx
     ln -s /usr/local/nginx/sbin/nginx /usr/sbin/nginx
     whereis nginx
@@ -57,10 +69,12 @@ function deploy_nginx()
 function deploy_shadowsocks()
 {
     cd
+    #install shadowsocks
     wget --no-check-certificate -O shadowsocksR.sh https://git.io/vHRHi
     chmod +x shadowsocksR.sh
     ./shadowsocksR.sh 2>&1 | tee shadowsocksR.log
     python /usr/local/shadowsocks/server.py -c /etc/shadowsocks.json -d start
+    #install bbr
     wget --no-check-certificate https://github.com/teddysun/across/raw/master/bbr.sh
     chmod +x bbr.sh
     ./bbr.sh
@@ -69,6 +83,17 @@ function deploy_shadowsocks()
 
 function deploy_python3()
 {
+    cd
+    #check existance of python3
+    check_str=$(whereis python3 | awk -F ': ' '{print $2}')
+    if [ "$check_str" != "" ];then
+        echo "python3 exist."
+        whereis python3
+        whereis pip3
+        python3 -V
+        return
+    fi
+    #start install python3
     yum install make gcc gcc-c++ 
     python --version
     wget -c https://www.python.org/ftp/python/3.6.4/Python-3.6.4.tgz
@@ -77,10 +102,20 @@ function deploy_python3()
     ./configure --prefix=/usr/local/python3 --enable-optimizations
     make
     make install
-    python -V
+    #create soft link
+    ln -s /usr/local/python3/bin/python3 /usr/bin/python3
+    ln -s /usr/local/python3/bin/python3 /usr/local/bin/python3
+    ln -s /usr/local/python3/bin/pip3 /usr/bin/pip3
+    ln -s /usr/local/python3/bin/pip3 /usr/local/bin/pip3
+    #show info
+    whereis python3
+    whereis pip3
     python3 -V
-
 }
+
+if [ $# -ne 1 ];then
+    usage
+fi
 env="$1"
 if [ "$env" == "work" ];then
     init_work
@@ -91,6 +126,5 @@ elif [ "$env" == "ss" ];then
 elif [ "$env" == "python3" ];then
     deploy_python3
 else
-    echo "usage: sh init_env.sh work/nginx/ss"
-    exit 0
+    usage
 fi
